pyproject.toml
[project]
name = "qemu-screenshot-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Vince Liem", email = "vincepaul.liem@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.25.0",
    "pillow>=12.1.0",
    "psutil>=7.2.1",
]

[project.scripts]
qemu-screenshot = "qemu_screenshot_mcp.server:main"

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"
